+++ app/controllers/open_api/v1/concerns/reservations_filters_concern.rb
@@ -45,6 +45,15 @@ module OpenAPI::V1::Concerns::ReservationsFiltersConcern
       reservations.where(reservable_id: may_array(filters[:reservable_id]))
     end
 
+    # @param reservations [ActiveRecord::Relation<Reservation>]
+    # @param filters [ActionController::Parameters]
+    def filter_by_availability_id(reservations, filters)
+      return reservations if filters[:availability_id].blank?
+
+      reservations.joins(:slots_reservations, :slots)
+                  .where(slots_reservations: { slots: { availability_id: may_array(filters[:availability_id]) } })
+    end
+
     # @param type [String]
     def format_type(type)
       type.singularize.classify