+++ app/controllers/open_api/v1/reservations_controller.rb
@@ -19,6 +19,7 @@ class OpenAPI::V1::ReservationsController < OpenAPI::V1::BaseController
     @reservations = filter_by_user(@reservations, params)
     @reservations = filter_by_reservable_type(@reservations, params)
     @reservations = filter_by_reservable_id(@reservations, params)
+    @reservations = filter_by_availability_id(@reservations, params)
 
     @reservations = @reservations.page(page).per(per_page)
     paginate @reservations, per_page: per_page